Implement Native share if available on cards + Copy link address in drop down + finish update/delete comment

+++ src/client/app/item/comment/comment.component.ts
@@ -1,8 +1,14 @@
-import { Component, Input, OnInit } from '@angular/core';
+import { Component, Input, OnInit, ViewChild } from '@angular/core';
+import { MatSnackBar } from '@angular/material';
 import { finalize } from 'rxjs/operators';
 import { CommentViewModel } from '../../../../shared/view-models/item/comment.view-model';
+import { ReportCommentViewModel } from '../../../../shared/view-models/item/report-comment.view-model';
+import { ContextMenuComponent } from '../../shared/context-menu/context-menu/context-menu.component';
+import { DialogService } from '../../shared/dialog/dialog.service';
 import { FormErrorsService } from '../../shared/form-errors/form-errors.service';
 import { AuthService } from '../../shared/services/auth.service';
+import { ShareService } from '../../shared/services/share.service';
+import { ShareDialogService } from '../../shared/share-dialog/share-dialog.service';
 import { ItemService } from '../item.service';
 
 @Component({
@@ -11,6 +17,7 @@ import { ItemService } from '../item.service';
   styleUrls: ['./comment.component.scss']
 })
 export class CommentComponent implements OnInit {
+  @ViewChild('contextMenu') contextMenu: ContextMenuComponent;
   @Input() comment: CommentViewModel;
   @Input() itemUserId: number;
   loggedInUserId = this.authService.getLoggedInUserId();
@@ -18,20 +25,75 @@ export class CommentComponent implements OnInit {
 
   constructor(private itemService: ItemService,
     private formErrorsService: FormErrorsService,
-    private authService: AuthService) { }
+    private authService: AuthService,
+    private dialogService: DialogService,
+    private shareDialogService: ShareDialogService,
+    private shareService: ShareService,
+    private snackBar: MatSnackBar) { }
 
   ngOnInit() {
   }
 
   deleteComment() {
-    this.isProcessing = true;
+    this.dialogService.confirm('Are you sure you want to delete this comment?').subscribe(data => {
+      if (data) {
+        this.contextMenu.close();
 
-    this.itemService.deleteComment(this.comment.uId)
-      .pipe(finalize(() => this.isProcessing = false))
-      .subscribe(data => {
-        console.log(data);
-      }, error => {
-        this.formErrorsService.updateFormValidity(error);
-      });
+        this.snackBar.open('Deleting...');
+
+        this.itemService.deleteComment(this.comment.uId)
+          .pipe(finalize(() => this.isProcessing = false))
+          .subscribe(() => {
+            this.snackBar.dismiss();
+            this.snackBar.open('Deleted');
+            // TODO: very dirty and bad UI but will work for now
+            location.reload();
+          }, error => {
+            this.snackBar.dismiss();
+            this.snackBar.open('Delete failed');
+            this.formErrorsService.updateFormValidity(error);
+          });
+      }
+    });
+  }
+
+  reportComment() {
+    this.dialogService.confirm('This comment is either spam, abusive, harmful or you think it doesn\'t belong on here.').subscribe(data => {
+      if (data) {
+        this.contextMenu.close();
+
+        const viewModel = new ReportCommentViewModel;
+        viewModel.uId = this.comment.uId;
+
+        this.snackBar.open('Sending...');
+
+        this.itemService.sendReport(viewModel)
+          .subscribe(() => {
+            this.snackBar.dismiss();
+            this.snackBar.open('Sent');
+          }, error => {
+            this.snackBar.dismiss();
+            this.snackBar.open('Sending failed');
+          });
+      }
+    });
+  }
+
+  openShareDialog() {
+    if (this.comment.itemUId) {
+      this.contextMenu.close();
+
+      const url = ['/item/comments', this.comment.itemUId];
+      if (!this.shareService.webShareWithUrl('Comment', url)) {
+        this.shareDialogService.share(url);
+      }
+    }
+  }
+
+  copyLink() {
+    if (this.comment.itemUId) {
+      this.shareService.copyWithUrl(['/item/comments', this.comment.itemUId]);
+      this.contextMenu.close();
+    }
   }
 }
